# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "江都 API",
    "version": "1.1.0",
    "description": "小程序+管理后台共用接口api"
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/fins/": {
      "post": {
        "tags": [
          "Fin"
        ],
        "summary": "获取Fin",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/fin_get"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Fin"
        ],
        "summary": "修改Fin",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/fin_update"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Fin"
        ],
        "summary": "删除Fin",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/fin_delete"
              }
            }
          }
        }
      }
    },
    "/api/fins/add": {
      "post": {
        "tags": [
          "Fin"
        ],
        "summary": "新建Fin",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/fin_add"
              }
            }
          }
        }
      }
    },
    "/api/fins/search": {
      "post": {
        "tags": [
          "Fin"
        ],
        "summary": "搜索Fin，query参数未知",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ]
      }
    },
    "/api/fins/finTags": {
      "post": {
        "tags": [
          "Fin"
        ],
        "summary": "获取标签",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/fin_finTags"
              }
            }
          }
        }
      }
    },
    "/api/fins/addTags": {
      "post": {
        "tags": [
          "Fin"
        ],
        "summary": "添加标签",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/fin_addTags"
              }
            }
          }
        }
      }
    },
    "/api/fins/deleteTags": {
      "post": {
        "tags": [
          "Fin"
        ],
        "summary": "删除标签",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/fin_deleteTags"
              }
            }
          }
        }
      }
    },
    "/api/fins/updateStatus": {
      "put": {
        "tags": [
          "Fin"
        ],
        "summary": "？？？？？？？",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "JiangduJWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/fin_updateStatus"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "fin_get": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "LOAN_TYPE": {
              "type": "string",
              "example": ""
            }
          }
        }
      }
    },
    "fin_add": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "FIN_ID": {
              "type": "string",
              "example": ""
            },
            "LOGO_URL": {
              "type": "string",
              "example": "1"
            },
            "FIN_CODE": {
              "type": "string",
              "example": ""
            },
            "NAME": {
              "type": "string",
              "example": "ddd"
            },
            "DESCRIPTION": {
              "type": "string",
              "example": "1"
            },
            "WORKFLOW_URL": {
              "type": "string",
              "example": "1"
            },
            "URL": {
              "type": "string",
              "example": "1"
            },
            "APPLY_URL": {
              "type": "string",
              "example": "1"
            },
            "BANK_NAME": {
              "type": "string",
              "example": ""
            },
            "BANK_ID": {
              "type": "string",
              "example": ""
            },
            "STATUS": {
              "type": "string",
              "example": "1"
            },
            "FIN_START_DATE": {
              "type": "string",
              "example": ""
            },
            "FIN_END_DATE": {
              "type": "string",
              "example": ""
            },
            "CREATED_AT": {
              "type": "string",
              "example": "2021-09-08T09:51:59.635Z"
            },
            "UPDATED_AT": {
              "type": "string",
              "example": "2021-09-08T09:51:59.635Z"
            },
            "USER_ID_USER_ID": {
              "type": "string",
              "example": "1"
            },
            "USER_NAME": {
              "type": "string",
              "example": "1"
            },
            "RATE_LOW": {
              "type": "string",
              "example": ""
            },
            "RATE_HIGH": {
              "type": "string",
              "example": ""
            },
            "LOAN_QUOTA": {
              "type": "string",
              "example": ""
            },
            "LOAN_DATE_LOW": {
              "type": "number",
              "example": 1
            },
            "LOAN_DATE_HIGH": {
              "type": "number",
              "example": 5
            },
            "CHARGE_METHOD": {
              "type": "string",
              "example": ""
            },
            "REPAY_METHOD": {
              "type": "string",
              "example": ""
            },
            "TARGET": {
              "type": "string",
              "example": ""
            },
            "FIN_TYPE": {
              "type": "string",
              "example": "1"
            },
            "PRODUCT_TYPE": {
              "type": "string",
              "example": ""
            },
            "LOAN_TYPE": {
              "type": "string",
              "example": ""
            }
          }
        },
        "token": {
          "type": "string",
          "example": "2c59a359-77a0-4b66-9d66-02089946a04c"
        }
      }
    },
    "fin_update": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "FIN_ID": {
              "type": "string",
              "example": "ac61b1b3-0c60-11ec-a492-90ccdf01187c"
            },
            "LOGO_URL": {
              "type": "string",
              "example": "null"
            },
            "FIN_CODE": {
              "type": "string",
              "example": "4080"
            },
            "NAME": {
              "type": "string",
              "example": "信用贷"
            },
            "DESCRIPTION": {
              "type": "string",
              "example": "这是一个介绍"
            },
            "WORKFLOW_URL": {
              "type": "string",
              "example": "null"
            },
            "URL": {
              "type": "string",
              "example": "null"
            },
            "APPLY_URL": {
              "type": "string",
              "example": "null"
            },
            "BANK_NAME": {
              "type": "string",
              "example": "中国银行"
            },
            "BANK_ID": {
              "type": "string",
              "example": "1234"
            },
            "STATUS": {
              "type": "string",
              "example": "启用"
            },
            "FIN_START_DATE": {
              "type": "string",
              "example": "2015-11-28 00:00:00.000000000"
            },
            "FIN_END_DATE": {
              "type": "string",
              "example": "2008-12-22 00:00:00.000000000"
            },
            "CREATED_AT": {
              "type": "string",
              "example": "2018-05-02"
            },
            "UPDATED_AT": {
              "type": "string",
              "example": "2003-12-13"
            },
            "USER_ID_USER_ID": {
              "type": "string",
              "example": "1234"
            },
            "USER_NAME": {
              "type": "string",
              "example": "null"
            },
            "RATE_LOW": {
              "type": "string",
              "example": "5%"
            },
            "RATE_HIGH": {
              "type": "string",
              "example": "9%"
            },
            "LOAN_QUOTA": {
              "type": "string",
              "example": "300w"
            },
            "LOAN_DATE_LOW": {
              "type": "number",
              "example": 6
            },
            "LOAN_DATE_HIGH": {
              "type": "number",
              "example": 26
            },
            "CHARGE_METHOD": {
              "type": "string",
              "example": "信用"
            },
            "REPAY_METHOD": {
              "type": "string",
              "example": "担保公司"
            },
            "TARGET": {
              "type": "string",
              "example": "小微企业"
            },
            "FIN_TYPE": {
              "type": "string",
              "example": "H"
            },
            "PRODUCT_TYPE": {
              "type": "string",
              "example": "线上审批"
            },
            "LOAN_TYPE": {
              "type": "string",
              "example": "流动资金贷款"
            }
          }
        }
      }
    },
    "fin_delete": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "text": {
              "type": "string",
              "example": "是否要删除启信宝(ID: ac61d8b6-0c60-11ec-aaff-90ccdf01187c)?"
            },
            "id": {
              "type": "string",
              "example": "ac61d8b6-0c60-11ec-aaff-90ccdf01187c"
            },
            "type": {
              "type": "string",
              "example": "fin"
            },
            "FIN_ID": {
              "type": "string",
              "example": "ac61d8b6-0c60-11ec-aaff-90ccdf01187c"
            }
          }
        }
      }
    },
    "fin_finTags": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "FIN_ID": {
              "type": "string",
              "example": "ac61b1b3-0c60-11ec-a492-90ccdf01187c"
            }
          }
        },
        "token": {
          "type": "string",
          "example": "2c59a359-77a0-4b66-9d66-02089946a04c"
        }
      }
    },
    "fin_addTags": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "FIN_ID_FIN_ID": {
              "type": "string",
              "example": "ac61b1b3-0c60-11ec-a492-90ccdf01187c"
            },
            "TAG_ID_TAG_ID": {
              "type": "string",
              "example": "15cefa4b-0c63-11ec-9ac5-342eb7e2a2ec"
            }
          }
        },
        "token": {
          "type": "string",
          "example": "2c59a359-77a0-4b66-9d66-02089946a04c"
        }
      }
    },
    "fin_deleteTags": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "FIN_ID_FIN_ID": {
              "type": "string",
              "example": "ac61b1b3-0c60-11ec-a492-90ccdf01187c"
            },
            "TAG_ID_TAG_ID": {
              "type": "string",
              "example": "15cefa4b-0c63-11ec-9ac5-342eb7e2a2ec"
            }
          }
        },
        "token": {
          "type": "string",
          "example": "2c59a359-77a0-4b66-9d66-02089946a04c"
        }
      }
    },
    "fin_updateStatus": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "COMMENTS": {
              "type": "string",
              "example": ""
            },
            "STATUS": {
              "type": "string",
              "example": ""
            },
            "USER_ID": {
              "type": "string",
              "example": ""
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JiangduJWT": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}